refactor: move components
containers/Loader/Factory/Steps/ApplyDevfile -> containers/Loader/Factory/Steps/Apply/Devfile
containers/Loader/Factory/Steps/ApplyResources -> containers/Loader/Factory/Steps/Apply/Resources

containers/Loader/Factory/Steps/FetchDevfile -> containers/Loader/Factory/Steps/Fetch/Devfile
containers/Loader/Factory/Steps/FetchResources -> containers/Loader/Factory/Steps/Fetch/Resources
