+++ packages/dashboard-frontend/src/containers/Loader/Factory/Steps/Apply/Resources/__mocks__/index.tsx
