+++ packages/dashboard-frontend/src/containers/Loader/Factory/Steps/Apply/Resources/__tests__/prepareResources.spec.ts
@@ -11,14 +11,14 @@
  */
 
 import { V1alpha2DevWorkspace } from '@devfile/api';
-import { DEVWORKSPACE_STORAGE_TYPE } from '../../../../../../services/devfileApi/devWorkspace/spec';
-import { generateWorkspaceName } from '../../../../../../services/helpers/generateName';
-import { DEVWORKSPACE_DEVFILE_SOURCE } from '../../../../../../services/workspace-client/devworkspace/devWorkspaceClient';
-import { DevWorkspaceResources } from '../../../../../../store/DevfileRegistries';
-import prepareResources from '../../../Steps/ApplyResources/prepareResources';
+import { DEVWORKSPACE_STORAGE_TYPE } from '../../../../../../../services/devfileApi/devWorkspace/spec';
+import { generateWorkspaceName } from '../../../../../../../services/helpers/generateName';
+import { DEVWORKSPACE_DEVFILE_SOURCE } from '../../../../../../../services/workspace-client/devworkspace/devWorkspaceClient';
+import { DevWorkspaceResources } from '../../../../../../../store/DevfileRegistries';
+import prepareResources from '../prepareResources';
 
 const suffix = '-1234';
-jest.mock('../../../../../../services/helpers/generateName.ts');
+jest.mock('../../../../../../../services/helpers/generateName.ts');
 (generateWorkspaceName as jest.Mock).mockImplementation(name => name + suffix);
 
 describe('FactoryLoaderContainer/prepareResources', () => {
