+++ packages/dashboard-frontend/src/containers/Loader/Factory/Steps/Fetch/Resources/__mocks__/index.tsx
